# docs/api-contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "ApiErrorPayload": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorPayload",
        "type": "object"
      },
...
      "ApiResponse_KnowledgeAcceptanceSnapshotData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeAcceptanceSnapshotData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[KnowledgeAcceptanceSnapshotData]",
        "type": "object"
      },
      "ApiResponse_KnowledgeImportResultData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeImportResultData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[KnowledgeImportResultData]",
        "type": "object"
      },
...
      "ApiResponse_PromptAcceptancePreviewData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PromptAcceptancePreviewData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[PromptAcceptancePreviewData]",
        "type": "object"
      },
...
      "ApiResponse_RuntimeActionData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeActionData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[RuntimeActionData]",
        "type": "object"
      },
...
      "ApiResponse_WebKnowledgeBuildResultData_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WebKnowledgeBuildResultData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[WebKnowledgeBuildResultData]",
        "type": "object"
      },
...
      "ApiResponse_list_KnowledgeSearchResultData__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/KnowledgeSearchResultData"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "success"
        ],
        "title": "ApiResponse[list[KnowledgeSearchResultData]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeAcceptanceSnapshotData": {
        "properties": {
          "imported_files": {
            "items": {
              "type": "string"
            },
            "title": "Imported Files",
            "type": "array"
          },
          "knowledge_status": {
            "$ref": "#/components/schemas/KnowledgeStatusData"
          },
          "retrieved_chunk_ids": {
            "items": {
              "type": "string"
            },
            "title": "Retrieved Chunk Ids",
            "type": "array"
          },
          "retrieved_chunks": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeSearchResultData"
            },
            "title": "Retrieved Chunks",
            "type": "array"
          },
          "search_query": {
            "default": "",
            "title": "Search Query",
            "type": "string"
          },
          "web_build_status": {
            "default": "",
            "title": "Web Build Status",
            "type": "string"
          }
        },
        "title": "KnowledgeAcceptanceSnapshotData",
        "type": "object"
      },
      "KnowledgeFileImportData": {
        "properties": {
          "file_name": {
            "default": "",
            "title": "File Name",
            "type": "string"
          },
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "KnowledgeFileImportData",
        "type": "object"
      },
      "KnowledgeImportRequest": {
        "additionalProperties": false,
        "properties": {
          "file_paths": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "File Paths",
            "type": "array"
          }
        },
        "required": [
          "file_paths"
        ],
        "title": "KnowledgeImportRequest",
        "type": "object"
      },
      "KnowledgeImportResultData": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeFileImportData"
            },
            "title": "Files",
            "type": "array"
          },
          "index_rebuilt": {
            "default": false,
            "title": "Index Rebuilt",
            "type": "boolean"
          }
        },
        "title": "KnowledgeImportResultData",
        "type": "object"
      },
      "KnowledgeSearchResultData": {
        "properties": {
          "chunk_id": {
            "default": "",
            "title": "Chunk Id",
            "type": "string"
          },
          "score": {
            "default": 0.0,
            "title": "Score",
            "type": "number"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          }
        },
        "title": "KnowledgeSearchResultData",
        "type": "object"
      },
...
      "PromptAcceptancePreviewData": {
        "properties": {
          "identity_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identity Confidence"
          },
          "identity_status": {
            "default": "",
            "title": "Identity Status",
            "type": "string"
          },
          "knowledge_results": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeSearchResultData"
            },
            "title": "Knowledge Results",
            "type": "array"
          },
          "latest_message": {
            "default": "",
            "title": "Latest Message",
            "type": "string"
          },
          "prompt_preview": {
            "default": "",
            "title": "Prompt Preview",
            "type": "string"
          },
          "resolved_user_id": {
            "default": "",
            "title": "Resolved User Id",
            "type": "string"
          },
          "scene": {
            "default": "friend",
            "title": "Scene",
            "type": "string"
          },
          "self_identity_profile": {
            "$ref": "#/components/schemas/PromptAcceptanceSelfIdentityData"
          }
        },
        "title": "PromptAcceptancePreviewData",
        "type": "object"
      },
...
      "RuntimeBootstrapStartRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "default": "global",
            "title": "Mode",
            "type": "string"
          },
          "narrator_settle_seconds": {
            "default": 10.0,
            "maximum": 60.0,
            "minimum": 0.0,
            "title": "Narrator Settle Seconds",
            "type": "number"
          },
          "poll_interval_seconds": {
            "default": 1.0,
            "maximum": 10.0,
            "minimum": 0.2,
            "title": "Poll Interval Seconds",
            "type": "number"
          },
          "ready_timeout_seconds": {
            "default": 20.0,
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Ready Timeout Seconds",
            "type": "number"
          },
          "wait_for_ui_ready_before_guardian": {
            "default": false,
            "title": "Wait For Ui Ready Before Guardian",
            "type": "boolean"
          }
        },
        "title": "RuntimeBootstrapStartRequest",
        "type": "object"
      },
      "RuntimeStartRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "default": "global",
            "title": "Mode",
            "type": "string"
          }
        },
        "title": "RuntimeStartRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebKnowledgeBuildRequest": {
        "additionalProperties": false,
        "properties": {
          "file_paths": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "File Paths",
            "type": "array"
          },
          "search_limit": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Search Limit",
            "type": "integer"
          }
        },
        "required": [
          "file_paths"
        ],
        "title": "WebKnowledgeBuildRequest",
        "type": "object"
      },
      "WebKnowledgeBuildResultData": {
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "title": "Documents",
            "type": "array"
          },
          "search_limit": {
            "default": 5,
            "title": "Search Limit",
            "type": "integer"
          },
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "WebKnowledgeBuildResultData",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WeChat AI Local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/debug/knowledge-acceptance": {
      "get": {
        "operationId": "knowledge_acceptance_api_v1_debug_knowledge_acceptance_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 8000,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "imported_files",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Imported Files"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeAcceptanceSnapshotData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Knowledge Acceptance",
        "tags": [
          "debug"
        ]
      }
    },
    "/api/v1/debug/prompt-preview": {
      "get": {
        "operationId": "prompt_preview_api_v1_debug_prompt_preview_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "latest_message",
            "required": true,
            "schema": {
              "maxLength": 8000,
              "minLength": 1,
              "title": "Latest Message",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "relationship_to_me",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Relationship To Me"
            }
          },
          {
            "in": "query",
            "name": "display_name",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 64,
              "title": "Display Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          },
          {
            "in": "query",
            "name": "knowledge_limit",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 10,
              "minimum": 1,
              "title": "Knowledge Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "scene",
            "required": false,
            "schema": {
              "default": "friend",
              "pattern": "^(friend|group)$",
              "title": "Scene",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PromptAcceptancePreviewData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prompt Preview",
        "tags": [
          "debug"
        ]
      }
    },
...
    "/api/v1/knowledge/import": {
      "post": {
        "operationId": "import_knowledge_api_v1_knowledge_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeImportResultData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/search": {
      "get": {
        "operationId": "search_knowledge_api_v1_knowledge_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_KnowledgeSearchResultData__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
...
    "/api/v1/knowledge/web-build": {
      "post": {
        "operationId": "build_web_knowledge_api_v1_knowledge_web_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebKnowledgeBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WebKnowledgeBuildResultData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Web Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
...
    "/api/v1/runtime/bootstrap-check": {
      "post": {
        "operationId": "runtime_bootstrap_check_api_v1_runtime_bootstrap_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeBootstrapStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RuntimeActionData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Bootstrap Check",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/v1/runtime/bootstrap-start": {
      "post": {
        "operationId": "runtime_bootstrap_start_api_v1_runtime_bootstrap_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeBootstrapStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RuntimeActionData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Bootstrap Start",
        "tags": [
          "runtime"
        ]
      }
    },
...
    "/api/v1/runtime/restart": {
      "post": {
        "operationId": "runtime_restart_api_v1_runtime_restart_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RuntimeActionData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Restart",
        "tags": [
          "runtime"
        ]
      }
    },
    "/api/v1/runtime/start": {
      "post": {
        "operationId": "runtime_start_api_v1_runtime_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RuntimeActionData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Start",
        "tags": [
          "runtime"
        ]
      }
    },
...
  }
}